+++ crates/common/src/python/mod.rs
@@ -27,6 +27,7 @@
 pub mod actor;
 pub mod cache;
 pub mod clock;
+pub mod component_msgbus;
 pub mod custom;
 pub mod enums;
 pub mod factory;